add main functionality of create event by group id

+++ backend/db/migrations/20230610173712-create-event.js
@@ -44,7 +44,7 @@ module.exports = {
         type: Sequelize.INTEGER
       },
       price: {
-        type: Sequelize.INTEGER
+        type: Sequelize.DECIMAL(10, 2)
       },
       startDate: {
         type: Sequelize.DATE,
